Combine Utils into already existing Helper class.
