+++ WeightBase.csproj
@@ -114,7 +114,6 @@
         <Compile Include="Properties\AssemblyInfo.cs" />
         <Compile Include="Tools\Helper.cs" />
         <Compile Include="Tools\ItemCache.cs" />
-        <Compile Include="Util.cs" />
         <Compile Include="VersionHandshake.cs" />
     </ItemGroup>
     <ItemGroup>
